feeds migrate-3048084-fix-tests-8
