+++ src/Plugin/migrate/process/Form/DefaultValueForm.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\feeds_migrate\Plugin\migrate\process\form;
+namespace Drupal\feeds_migrate\Plugin\migrate\process\Form;
 
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Form\FormStateInterface;
