+++ src/Plugin/migrate/source/Form/SourceFormPluginBase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\feeds_migrate\Plugin\migrate\source\form;
+namespace Drupal\feeds_migrate\Plugin\migrate\source\Form;
 
 use Drupal\feeds_migrate\Plugin\MigrateFormPluginBase;
