+++ src/Plugin/migrate_plus/data_fetcher/Form/HttpForm.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\feeds_migrate\Plugin\migrate_plus\data_fetcher\form;
+namespace Drupal\feeds_migrate\Plugin\migrate_plus\data_fetcher\Form;
 
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Form\FormStateInterface;
@@ -9,14 +9,14 @@ use Drupal\Core\Form\FormStateInterface;
  * The configuration form for the file migrate data fetcher plugin.
  *
  * @MigrateForm(
- *   id = "file",
- *   title = @Translation("File Data Fetcher Plugin Form"),
+ *   id = "http",
+ *   title = @Translation("Http Fetcher Plugin Form"),
  *   form = "configuration",
- *   parent_id = "file",
+ *   parent_id = "http",
  *   parent_type = "data_fetcher"
  * )
  */
-class FileForm extends DataFetcherFormPluginBase {
+class HttpForm extends DataFetcherFormPluginBase {
 
   /**
    * {@inheritdoc}
@@ -26,11 +26,10 @@ class FileForm extends DataFetcherFormPluginBase {
     $entity = $this->entity;
     $source = $entity->get('source');
 
-    $form['directory'] = [
+    $form['urls'] = [
       '#type' => 'textfield',
-      '#title' => $this->t('File Upload Directory'),
-      // @todo move this to defaultConfiguration
-      '#default_value' => $source['data_fetcher']['directory'] ?: 'public://migrate',
+      '#title' => $this->t('File Location (single location only)'),
+      '#default_value' => $source['urls'] ?: 'public://migrate/data.xml',
     ];
 
     return $form;
@@ -40,7 +39,8 @@ class FileForm extends DataFetcherFormPluginBase {
    * {@inheritdoc}
    */
   public function copyFormValuesToEntity(EntityInterface $entity, array $form, FormStateInterface $form_state) {
-    $entity->source['data_fetcher']['directory'] = $form_state->getValue('directory');
+    // The urls property goes right into source, not under data_fetcher.
+    $entity->source['urls'] = $form_state->getValue('urls');
   }
 
 }
